requirements.txt
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
ConfigArgParse==1.5.3
cycler==0.11.0
DBUtils==3.0.2
elastic-transport==8.4.0
elasticsearch==8.5.0
Faker==13.15.1
Flask==2.1.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
fonttools==4.37.1
gevent==21.12.0
geventhttpclient==2.0
gmssl==3.2.1
google-auth==2.11.0
greenlet==1.1.2
grpcio==1.47.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
kubernetes==24.2.0
locust==2.10.1
MarkupSafe==2.1.1
matplotlib==3.5.3
msgpack==1.0.4
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.3
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==22.3.0
requests==2.28.1
requests-oauthlib==1.3.1
roundrobin==0.0.2
rsa==4.9
schedule==1.1.0
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.11
websocket-client==1.4.0
Werkzeug==2.2.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0